requirements.txt
faker==22.0.0
pymongo==4.6.0
python-dotenv==1.0.0
flask==3.0.0
waitress==2.1.2